Split backup control plane tests by module
